refactor codebase to modular architecture
