+++ src/middleware/validation.ts
@@ -2,7 +2,7 @@ import { Request, Response, NextFunction } from 'express';
 import { validate } from 'class-validator';
 import { BadRequestError } from '../middleware/error';
 import { ClassConstructor, plainToClass, plainToInstance } from 'class-transformer';
-import logger from '../utils/winston';
+import logger from '../modules/logging/winston';
 
 // a middleware to validate the user input
 export default class RequestValidator {
