+++ src/modules/auth/auth.interface.ts
