+++ src/modules/cloud/cloudinary.interface.ts
