+++ src/modules/event/event.contoller.ts
@@ -1,15 +1,16 @@
 import {Response} from "express";
 import { AuthRequest } from "../interfaces/userAuth";
 import { prisma } from "../utils/db";
-import {BadRequestError, ConflictError, NotAuthorizedError, NotFoundError } from "../middleware/error";
+import { AuthError, BadRequestError, NotAuthorizedError, NotFoundError } from "../middleware/error";
 import MailService from "../service/mailing";
-import logger from "../utils/winston";
+import logger from "../modules/logging/winston";
 import cloudinaryInstance from "../service/cloudinary";
 
 
 
 
 export default class EventController {
+
 
     static createEvent = async (req:AuthRequest,  res:Response) => {
         let {name, description, date, startTime, endTime,  location, totalCapacity} = req.body;
@@ -48,7 +49,7 @@ export default class EventController {
         const event = await prisma.event.findUnique({
             where:{ id: eventId }  });
            
-        if(!event) { throw new NotFoundError("No event with Id!") }
+        if(!event) { throw new BadRequestError("No event with Id!") }
         
         // Check if event is at capacity before adding the new attendee
 
@@ -132,6 +133,7 @@ export default class EventController {
     }
 
     static udpateEvent = async (req:AuthRequest,  res:Response) => {
+
         // check if current event belongs to signed in User.
         let event = await prisma.event.findUnique({
             where:{
@@ -139,8 +141,8 @@ export default class EventController {
             }
         })
 
-        if(!event) {throw new ConflictError()}
-        if(event.creatorId != req.user.id) {throw new NotAuthorizedError();}
+        if(!event) throw new NotFoundError("No Event with Id Found!")
+        if(event.creatorId != req.user.id) throw new NotAuthorizedError();
 
         let {name, description, date, startTime, endTime,  location, totalCapacity} = req.body;
         date = new Date(date);
@@ -148,10 +150,10 @@ export default class EventController {
 
         event = await prisma.event.update({
             where:{
-                id: req.params.id
+                id: req.user.id
             },
             data:{
-                name, description, date, startTime, endTime, totalCapacity
+                name, description, date, startTime, endTime, location, totalCapacity
             }
         })
