+++ src/modules/event/event.interface.ts
@@ -1,0 +1,14 @@
+import { Request } from 'express';
+
+export interface AuthRequest extends Request {
+  user?: any;
+}
+
+
+export interface User {
+  firstname: string,
+  lastname: string,
+  email: string,
+  password: string,
+  contact: string
+}
