+++ src/modules/mailing/mail.interface.ts
